requirements.txt
openai

python-dotenv

langchain-openai

langchain_google_genai

langchain-anthropic

langchain-ollama

langchain_core

streamlit

gradio

huggingface_hub

langchain_huggingface

transformers

torch

langchain

langchain-aws

groq

langchain-groq

langchain-community

pypdf

bs4